se agrega los scripts para subir items con imagenes y el delete se arregla

+++ index.js
@@ -7,32 +7,28 @@ import colors from 'colors'; // Import colors package
 async function main() {
 
 
- try {
-   const inicio = process.hrtime();
-
-// Tu código aquí
-// Incluye llamadas asíncronas si es necesario
-
-
-
+  try {
+    
+    const inicio = process.hrtime();
 
     const loginController = new LoginController();
     const sesionCookie = await loginController.loginController();
-    // const sesionCookie    = '';
-        const importController = new ImportController();
-     //await importController.importController(sesionCookie); 
-        //
-        const deleteService = new DeleteService();
-      // 
-      await deleteService.deleteItem(sesionCookie);
-   const fin = process.hrtime(inicio);
-   const tiempoDeEjecucionEnNanosegundos = fin[0] * 1e9 + fin[1];
-   const tiempoDeEjecucionEnSegundos = tiempoDeEjecucionEnNanosegundos / 1e9;
 
-   console.log(colors.yellow(`El tiempo de ejecución fue de ${tiempoDeEjecucionEnSegundos} segundos.`));
- } catch (error) {
-console.log(error)    
- }    
+    if (process.argv[2]==='delete') {
+      const deleteService = new DeleteService();
+      await deleteService.deleteItem(sesionCookie);
+      
+    }else{
+      const importController = new ImportController();
+        await importController.importController(sesionCookie);
+    }
+    const fin = process.hrtime(inicio);
+    const tiempoDeEjecucionEnNanosegundos = fin[0] * 1e9 + fin[1];
+    const tiempoDeEjecucionEnSegundos = tiempoDeEjecucionEnNanosegundos / 1e9;
+    console.log(colors.yellow(`El tiempo de ejecución fue de ${tiempoDeEjecucionEnSegundos} segundos.`));
+  } catch (error) {
+    console.log(error)
+  }
 }
 
 await main();
